requirements.txt
numpy
pandas
pymongo
python-dotenv
fastapi
uvicorn

# -e .